Added minor refactor to channel deletion + now tito is added as an user every time a new channel is created
