+++ routes/messages.js
@@ -121,7 +121,6 @@ async function sendMessageToTopic(sender, workspace, channel, message) {
     },
     topic: topic
   };
-  console.log(fbMessage);
   await firebase.sendMessageToTopic(fbMessage);
 }
